+++ owner/src/main/java/org/the_chance/honeymart/ui/components/ContentVisibility.kt
@@ -1,4 +1,4 @@
-package org.the_chance.honeymart.ui.composables
+package org.the_chance.honeymart.ui.components
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.AnimatedVisibilityScope
