+++ owner/src/main/java/org/the_chance/honeymart/ui/features/category/CategoriesScreen.kt
@@ -19,27 +19,29 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
-import org.the_chance.honeymart.ui.features.add_product.components.AddProductContent
-import org.the_chance.honeymart.ui.features.category.composable.AddCategoryContent
-import org.the_chance.honeymart.ui.features.category.composable.CategoryItems
-import org.the_chance.honeymart.ui.features.category.composable.CategoryProducts
+import org.the_chance.design_system.R
 import org.the_chance.honeymart.ui.features.category.composable.EmptyCategory
 import org.the_chance.honeymart.ui.features.category.composable.HoneyMartTitle
-import org.the_chance.honeymart.ui.features.category.composable.UpdateCategoryContent
-import org.the_chance.honeymart.ui.features.product_details.composables.ProductDetailsContent
+import org.the_chance.honeymart.ui.features.category.content.AddCategoryContent
+import org.the_chance.honeymart.ui.features.category.content.AddProductContent
+import org.the_chance.honeymart.ui.features.category.content.CategoryItemsContent
+import org.the_chance.honeymart.ui.features.category.content.CategoryProductsContent
+import org.the_chance.honeymart.ui.features.category.content.ProductDetailsContent
+import org.the_chance.honeymart.ui.features.category.content.UpdateCategoryContent
 import org.the_chance.honymart.ui.composables.ConnectionErrorPlaceholder
 import org.the_chance.honymart.ui.composables.CustomAlertDialog
 import org.the_chance.honymart.ui.composables.Loading
 import org.the_chance.honymart.ui.composables.SnackBarWithDuration
-import org.the_chance.design_system.R
 
 /**
  * Created by Aziza Helmy on 8/7/2023.
  */
+
 @Composable
 fun CategoriesScreen(categoriesViewModel: CategoriesViewModel = hiltViewModel()) {
     val lifecycleOwner = LocalLifecycleOwner.current
     val state by categoriesViewModel.state.collectAsState()
+
     CategoriesContent(state, categoriesViewModel)
     LaunchedEffect(lifecycleOwner) {
         categoriesViewModel.getAllCategory()
@@ -51,14 +53,12 @@ fun CategoriesContent(
     state: CategoriesUiState,
     listener: CategoriesInteractionsListener,
 ) {
-
     Column(
         modifier = Modifier
             .fillMaxSize()
             .background(MaterialTheme.colorScheme.tertiaryContainer)
     ) {
         HoneyMartTitle()
-
         Loading(state = state.isLoading && state.categories.isEmpty())
         AnimatedVisibility(visible = state.placeHolderCondition()) {
             Column(
@@ -72,7 +72,6 @@ fun CategoriesContent(
                 )
             }
         }
-
 
         Row(modifier = Modifier.fillMaxSize()) {
             Loading(state = state.isLoading && state.categories.isNotEmpty())
@@ -90,59 +89,57 @@ fun CategoriesContent(
                     state = state.categories.isEmpty() && !state.isLoading && !state.isError,
                     onClick = { listener.resetShowState(Visibility.ADD_CATEGORY) }
                 )
+                AnimatedVisibility(
+                    visible = state.categories.isNotEmpty() && state.showScreenState.showFab
+                ) {
+                    CategoryItemsContent(state = state, listener = listener)
+                }
 
-                CategoryItems(state = state, listener = listener)
-
-                AnimatedVisibility(visible = state.showCategoryProductsInCategory())
-                { CategoryProducts(state = state, listener = listener) }
-
+                AnimatedVisibility(visible = state.showCategoryProductsInCategory()) {
+                    CategoryProductsContent(state = state, listener = listener)
+                }
             }
             Column(
                 modifier = Modifier
                     .fillMaxSize()
                     .weight(1f)
             ) {
-                AnimatedVisibility(visible = state.showScreenState.showAddCategory)
-                {
+                AnimatedVisibility(visible = state.showScreenState.showAddCategory) {
                     AddCategoryContent(listener = listener, state = state)
                 }
 
-                AnimatedVisibility(visible = state.showCategoryProductsInProduct())
-                {
-                    CategoryProducts(state = state, listener = listener)
+                AnimatedVisibility(visible = state.showCategoryProductsInProduct()) {
+                    CategoryProductsContent(state = state, listener = listener)
                 }
 
-                AnimatedVisibility(visible = state.showScreenState.showUpdateCategory)
-                {
+                AnimatedVisibility(visible = state.showScreenState.showUpdateCategory) {
                     UpdateCategoryContent(state = state, listener = listener)
                 }
 
-                AnimatedVisibility(visible = state.showAddProductContent())
-                {
+                AnimatedVisibility(visible = state.showAddProductContent()) {
                     AddProductContent(state = state, listener = listener)
                 }
-                AnimatedVisibility(visible = state.showProductDetailsContent())
-                {
+
+                AnimatedVisibility(visible = state.showProductDetailsContent()) {
                     ProductDetailsContent(
                         titleScreen = stringResource(id = R.string.product_details),
                         confirmButton = stringResource(id = R.string.update),
                         cancelButton = stringResource(id = R.string.delete),
                         state = state,
                         listener = listener,
-                        onClickConfirm = {listener.onClickUpdateProductDetails() },
+                        onClickConfirm = { listener.onClickUpdateProductDetails() },
                         onClickCancel = { listener.resetShowState(Visibility.DELETE_PRODUCT) }
                     )
                 }
-                AnimatedVisibility(visible = state.showProductUpdateContent())
-                {
+                AnimatedVisibility(visible = state.showProductUpdateContent()) {
                     ProductDetailsContent(
                         titleScreen = stringResource(id = R.string.update_product),
                         confirmButton = stringResource(id = R.string.save),
                         cancelButton = stringResource(id = R.string.cancel),
                         state = state,
                         listener = listener,
-                        onClickConfirm = {listener.updateProductDetails(state)},
-                        onClickCancel = {listener.onClickCancel() }
+                        onClickConfirm = { listener.updateProductDetails(state) },
+                        onClickCancel = { listener.onClickCancel() }
                     )
                 }
             }
@@ -165,7 +162,6 @@ fun CategoriesContent(
         onClickTryAgain = listener::getAllCategory
     )
 
-
     if (state.showScreenState.showDialog) {
         CustomAlertDialog(
             message = stringResource(R.string.you_delete_a_categories) +
@@ -182,16 +178,13 @@ fun CategoriesContent(
 
     if (state.showScreenState.showDeleteDialog) {
         CustomAlertDialog(
-            message = stringResource(R.string.you_delete_a_product) +
-                    stringResource(R.string.are_you_sure),
+            message = stringResource(R.string.you_delete_a_product) + stringResource(R.string.are_you_sure),
             onConfirm = {
                 listener.deleteProductById(state.newProducts.id)
                 listener.resetShowState(Visibility.DELETE_PRODUCT)
             },
-
             onCancel = { listener.resetShowState(Visibility.DELETE_PRODUCT) },
             onDismissRequest = { listener.resetShowState(Visibility.DELETE_PRODUCT) }
         )
     }
 }
-
