+++ owner/src/main/java/org/the_chance/honeymart/ui/features/category/composable/AddImageButton.kt
@@ -1,4 +1,4 @@
-package org.the_chance.honeymart.ui.features.add_product.components
+package org.the_chance.honeymart.ui.features.category.composable
 
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.ActivityResultLauncher
