+++ owner/src/main/java/org/the_chance/honeymart/ui/features/category/composable/DropDownMenu.kt
@@ -1,8 +1,7 @@
-package org.the_chance.honeymart.ui.components
+package org.the_chance.honeymart.ui.features.category.composable
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.material.DropdownMenu
 import androidx.compose.material.icons.Icons
@@ -20,11 +19,9 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import org.the_chance.honymart.ui.theme.black8
 
 @Composable
 fun DropDownMenuList(
