+++ owner/src/main/java/org/the_chance/honeymart/ui/features/login/LoginScreen.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
 import org.the_chance.honeymart.LocalNavigationProvider
-import org.the_chance.honeymart.ui.composables.HoneyAuthScaffold
+import org.the_chance.honeymart.ui.components.HoneyAuthScaffold
 import org.the_chance.honeymart.ui.features.category.navigateToCategoryScreen
 import org.the_chance.honeymart.ui.features.signup.navigateToSignupScreen
 import org.the_chance.honymart.ui.composables.HoneyAuthFooter
@@ -113,7 +113,7 @@ fun LoginContent(
                 text = stringResource(R.string.don_t_have_an_account),
                 textButtonText = stringResource(R.string.Sign_up),
                 onTextButtonClicked = listener::onClickSignup,
-                modifier = Modifier.Companion.align(Alignment.CenterHorizontally)
+                modifier = Modifier.align(Alignment.CenterHorizontally)
             )
         }
     }
