--- owner/src/main/java/org/the_chance/honeymart/ui/features/update_category/UpdateCategoryInteractionListener.kt
@@ -1,9 +1,0 @@
-package org.the_chance.honeymart.ui.features.update_category
-
-interface UpdateCategoryInteractionListener {
-
-    fun onUpdatedCategoryNameChanged(name: String)
-    fun onClickCategoryIcon(categoryIconId: Int)
-    fun updateCategory(category: UpdateCategoryUiState)
-    fun onClickCancelUpdateCategory()
-}
